# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Actions Runbook API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RunbookSeverity = typeof RunbookSeverity[keyof typeof RunbookSeverity];


export const RunbookSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Runbook {
  id: number;
  title: string;
  system: string;
  severity: RunbookSeverity;
  steps: string;
  rollback: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export type RunbookInputSeverity = typeof RunbookInputSeverity[keyof typeof RunbookInputSeverity];


export const RunbookInputSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface RunbookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  system: string;
  severity: RunbookInputSeverity;
  steps: string;
  rollback: string;
  tags?: string[];
}

export type RunbookUpdateSeverity = typeof RunbookUpdateSeverity[keyof typeof RunbookUpdateSeverity];


export const RunbookUpdateSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface RunbookUpdate {
  /** @minLength 1 */
  title?: string;
  system?: string;
  severity?: RunbookUpdateSeverity;
  steps?: string;
  rollback?: string;
  tags?: string[];
}

export type RunbookStatsBySeverity = {[key: string]: number};

export type RunbookStatsBySystem = {[key: string]: number};

export interface RunbookStats {
  total: number;
  bySeverity: RunbookStatsBySeverity;
  bySystem: RunbookStatsBySystem;
}

export interface SeedResult {
  inserted: number;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListRunbooksParams = {
/**
 * Full-text search across title, system, steps
 */
search?: string;
/**
 * Filter by system name
 */
system?: string;
/**
 * Filter by severity level
 */
severity?: ListRunbooksSeverity;
/**
 * Filter by tag
 */
tag?: string;
};

export type ListRunbooksSeverity = typeof ListRunbooksSeverity[keyof typeof ListRunbooksSeverity];


export const ListRunbooksSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;