# lib/api-zod/src/generated/types/runbook.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Actions Runbook API
 * OpenAPI spec version: 0.1.0
 */
import type { RunbookSeverity } from './runbookSeverity';

export interface Runbook {
  id: number;
  title: string;
  system: string;
  severity: RunbookSeverity;
  steps: string;
  rollback: string;
  tags: string[];
  createdAt: Date;
  updatedAt: Date;
}