# lib/api-zod/src/generated/types/runbookInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Actions Runbook API
 * OpenAPI spec version: 0.1.0
 */
import type { RunbookInputSeverity } from './runbookInputSeverity';

export interface RunbookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  system: string;
  severity: RunbookInputSeverity;
  steps: string;
  rollback: string;
  tags?: string[];
}