# lib/api-zod/src/generated/types/runbookStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Actions Runbook API
 * OpenAPI spec version: 0.1.0
 */
import type { RunbookStatsBySeverity } from './runbookStatsBySeverity';
import type { RunbookStatsBySystem } from './runbookStatsBySystem';

export interface RunbookStats {
  total: number;
  bySeverity: RunbookStatsBySeverity;
  bySystem: RunbookStatsBySystem;
}