# lib/api-zod/src/generated/types/runbookUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Actions Runbook API
 * OpenAPI spec version: 0.1.0
 */
import type { RunbookUpdateSeverity } from './runbookUpdateSeverity';

export interface RunbookUpdate {
  /** @minLength 1 */
  title?: string;
  system?: string;
  severity?: RunbookUpdateSeverity;
  steps?: string;
  rollback?: string;
  tags?: string[];
}