# lib/api-zod/src/generated/types/seedResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GitHub Actions Runbook API
 * OpenAPI spec version: 0.1.0
 */

export interface SeedResult {
  inserted: number;
  message: string;
}